focus_news/requirements.txt
schedule>=1.1.0
python-decouple>=3.5

pandas>=2.2.3

Telethon

SQLAlchemy>=1.4.26
pymongo
psycopg2-binary

openai>=1.50.2
requests>=2.32.3

playwright>1.4.3